% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankstability.R
\name{bankstability}
\alias{bankstability}
\title{Bank stability metrics}
\usage{
bankstability(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Bank stability metrics
}
\examples{
bankstability(sampdat)
}
