% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelmorph.R
\name{channelmorph}
\alias{channelmorph}
\title{Channel morphology metrics}
\usage{
channelmorph(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Channel morphology metrics
}
\examples{
channelmorph(sampdat)
}
