% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelsinuosity.R
\name{channelsinuosity}
\alias{channelsinuosity}
\title{Channel sinuosity metrics}
\usage{
channelsinuosity(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Channel sinuosity metrics
}
\examples{
channelsinuosity(sampdat)
}
