% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densiometer.R
\name{densiometer}
\alias{densiometer}
\title{Densiometer metrics}
\usage{
densiometer(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Densiometer metrics
}
\examples{
sampdat <- phabformat(sampdat)
densiometer(sampdat)
}
