% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long-format.R
\name{long_form}
\alias{long_form}
\title{Format output in long form}
\usage{
long_form(out)
}
\arguments{
\item{out}{}
}
\value{
result data frame in long format
}
\description{
Format output in long form
}
\examples{
\dontrun{
res <- long_form(out)
}
}
