% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longformat.R
\name{longformat}
\alias{longformat}
\title{Melt the phabmetrics output dataframe to long format}
\usage{
longformat(metrics)
}
\arguments{
\item{data}{Input data}
}
\description{
Melt the phabmetrics output dataframe to long format
}
