% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampdat.R
\docType{data}
\name{sampdat}
\alias{sampdat}
\title{Sample data}
\format{A \code{\link{data.frame}} in long format, one row per sample}
\usage{
sampdat
}
\description{
Sample data for calculating PHAB metrics
}
\examples{
data(sampdat)
}
\keyword{datasets}
