% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phabmetrics.R
\name{phabmetrics}
\alias{phabmetrics}
\title{Calculate all PHAB metrics}
\usage{
phabmetrics(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Calculate all PHAB metrics
}
\examples{
\dontrun{
phabmetrics(sampdat)
}
}
